backend/requirements.txt
fastapi
pydantic
openai
qdrant-client
langchain-text-splitters
sentence-transformers
python-dotenv
pypdf2
python-multipart